+++ src/main/java/ecdar/controllers/DeclarationsController.java
@@ -41,20 +41,20 @@ public class DeclarationsController implements Initializable {
      */
     private void initializeWidthAndHeight() {
         // Fetch width and height of canvas and update
-        root.setMinWidth(EcdarController.getActiveCanvasPresentation().getController().getWidthProperty().doubleValue());
-        canvasHeight = EcdarController.getActiveCanvasPresentation().getController().getHeightProperty().doubleValue();
-        updateOffset(EcdarController.getActiveCanvasPresentation().getController().getInsetShouldShow().get());
+        root.setMinWidth(MainController.getActiveCanvasPresentation().getController().getWidthProperty().doubleValue());
+        canvasHeight = MainController.getActiveCanvasPresentation().getController().getHeightProperty().doubleValue();
+        updateOffset(MainController.getActiveCanvasPresentation().getController().getInsetShouldShow().get());
         updateHeight();
 
-        EcdarController.getActiveCanvasPresentation().getController().getWidthProperty().addListener((observable, oldValue, newValue) -> {
+        MainController.getActiveCanvasPresentation().getController().getWidthProperty().addListener((observable, oldValue, newValue) -> {
             root.setMinWidth(newValue.doubleValue());
             root.setMaxWidth(newValue.doubleValue());
         });
-        EcdarController.getActiveCanvasPresentation().getController().getHeightProperty().addListener((observable, oldValue, newValue) -> {
+        MainController.getActiveCanvasPresentation().getController().getHeightProperty().addListener((observable, oldValue, newValue) -> {
             canvasHeight = newValue.doubleValue();
             updateHeight();
         });
-        EcdarController.getActiveCanvasPresentation().getController().getInsetShouldShow().addListener((observable, oldValue, newValue) -> {
+        MainController.getActiveCanvasPresentation().getController().getInsetShouldShow().addListener((observable, oldValue, newValue) -> {
             updateOffset(newValue);
             updateHeight();
         });
@@ -102,7 +102,7 @@ public class DeclarationsController implements Initializable {
      * Updates the height of the view.
      */
     private void updateHeight() {
-        final double value = canvasHeight - EcdarController.getActiveCanvasPresentation().getController().DECLARATION_Y_MARGIN - offSet;
+        final double value = canvasHeight - MainController.getActiveCanvasPresentation().getController().DECLARATION_Y_MARGIN - offSet;
 
         root.setMinHeight(value);
         root.setMaxHeight(value);
