+++ src/main/java/ecdar/controllers/EcdarController.java
@@ -66,25 +66,17 @@ public class EcdarController implements Initializable, Presentable {
 
     // View stuff
     public StackPane root;
-    public SimulatorPresentation simulatorPresentation = new SimulatorPresentation();
     public QueryPanePresentation queryPane;
     public ProjectPanePresentation filePane;
     public StackPane toolbar;
     public Label queryPaneFillerElement;
     public Label filePaneFillerElement;
-    public StackPane dialogContainer;
-    public JFXDialog dialog;
+    public Rectangle bottomFillerElement;
     public StackPane modalBar;
-    public JFXTextField queryTextField;
-    public JFXTextField commentTextField;
     public JFXRippler colorSelected;
     public JFXRippler deleteSelected;
     public JFXRippler undo;
     public JFXRippler redo;
-    public JFXTabPane tabPane;
-    public Tab errorsTab;
-    public Tab warningsTab;
-    public Rectangle tabPaneResizeElement;
     public StackPane tabPaneContainer;
 
     public ImageView helpInitialImage;
@@ -100,16 +92,6 @@ public class EcdarController implements Initializable, Presentable {
     public JFXButton switchToOutputButton;
     public JFXToggleButton switchEdgeStatusButton;
 
-    public MenuItem menuEditMoveLeft;
-    public MenuItem menuEditMoveUp;
-    public MenuItem menuEditMoveRight;
-    public MenuItem menuEditMoveDown;
-    public JFXButton testHelpAcceptButton;
-    public JFXDialog testHelpDialog;
-    public StackPane testHelpContainer;
-    public StackPane aboutContainer;
-    public JFXDialog aboutDialog;
-    public JFXButton aboutAcceptButton;
     public StackPane canvasPane;
 
     private double expandHeight = 300;
@@ -125,65 +107,6 @@ public class EcdarController implements Initializable, Presentable {
             setMaxHeight(35 + frac * (expandHeight - 35));
         }
     };
-    public Rectangle bottomFillerElement;
-    public JFXRippler collapseMessages;
-    public FontIcon collapseMessagesIcon;
-    public ScrollPane errorsScrollPane;
-    public VBox errorsList;
-    public ScrollPane warningsScrollPane;
-    public VBox warningsList;
-    public Tab backendErrorsTab;
-    public ScrollPane backendErrorsScrollPane;
-    public VBox backendErrorsList;
-    public HBox centerContainer;
-
-    // The program top menu
-    public MenuBar menuBar;
-    public MenuItem menuBarViewFilePanel;
-    public MenuItem menuBarViewQueryPanel;
-    public MenuItem menuBarViewGrid;
-    public MenuItem menuBarViewCanvasSplit;
-    public MenuItem menuBarViewSimLeftPanel;
-    public MenuItem menuBarViewSimRightPanel;
-    public MenuItem menuBarViewHome;
-    public MenuItem menuBarViewSim;
-    public MenuItem menuBarFileCreateNewProject;
-    public MenuItem menuBarFileOpenProject;
-    public MenuItem menuBarFileSave;
-    public MenuItem menuBarFileSaveAs;
-    public MenuItem menuBarFileNewMutationTestObject;
-    public MenuItem menuBarFileExportAsPng;
-    public MenuItem menuBarFileExportAsPngNoBorder;
-    public MenuItem menuBarZoomInSimulator;
-    public MenuItem menuBarZoomOutSimulator;
-    public MenuItem menuBarZoomResetSimulator;
-    public MenuItem menuBarOptionsCache;
-    public MenuItem menuBarHelpHelp;
-    public MenuItem menuBarHelpAbout;
-    public MenuItem menuBarHelpTest;
-
-    public JFXSnackbar snackbar;
-    public HBox statusBar;
-    public Label statusLabel;
-    public Label queryLabel;
-    public HBox queryStatusContainer;
-
-    public StackPane queryDialogContainer;
-    public JFXDialog queryDialog;
-    public Text queryTextResult;
-    public Text queryTextQuery;
-
-    private static JFXDialog _queryDialog;
-    private static Text _queryTextResult;
-    private static Text _queryTextQuery;
-    public JFXDialog reloadSimulatorDialog;
-    private static JFXDialog _reloadSimulatorDialog;
-
-    private double tabPanePreviousY = 0;
-    public boolean shouldISkipOpeningTheMessagesContainer = true;
-
-    private static final ObjectProperty<CanvasPresentation> activeCanvasPresentation = new SimpleObjectProperty<>(new CanvasPresentation());
-    private static final ObjectProperty<EcdarController.Mode> currentMode = new SimpleObjectProperty<>(EcdarController.Mode.Editor);
 
     public static void runReachabilityAnalysis() {
         if (!reachabilityServiceEnabled) return;
@@ -197,54 +120,9 @@ public class EcdarController implements Initializable, Presentable {
 
     @Override
     public void initialize(final URL location, final ResourceBundle resources) {
-        dialog.setDialogContainer(dialogContainer);
-        dialogContainer.opacityProperty().bind(dialog.getChildren().get(0).scaleXProperty());
-        dialog.setOnDialogClosed(event -> dialogContainer.setVisible(false));
-
-        _queryDialog = queryDialog;
-        _queryTextResult = queryTextResult;
-        _queryTextQuery = queryTextQuery;
-        queryDialog.setDialogContainer(queryDialogContainer);
-        queryDialogContainer.opacityProperty().bind(queryDialog.getChildren().get(0).scaleXProperty());
-        queryDialog.setOnDialogClosed(event -> {
-            queryDialogContainer.setVisible(false);
-            queryDialogContainer.setMouseTransparent(true);
-        });
-        queryDialog.setOnDialogOpened(event -> {
-            queryDialogContainer.setVisible(true);
-            queryDialogContainer.setMouseTransparent(false);
-        });
-
-        initializeCanvasPane();
-
         initializeEdgeStatusHandling();
-
-        this.initializeMode();
         initializeKeybindings();
-        initializeTabPane();
-        initializeStatusBar();
-        initializeMessages();
-        initializeMenuBar();
         initializeReachabilityAnalysisThread();
-        this.simulatorPresentation.getController().root.prefWidthProperty().bind(this.centerContainer.widthProperty());
-
-    }
-
-    private void initializeMode() {
-        currentMode.addListener((obs, oldMode, newMode) -> {
-            if (newMode == EcdarController.Mode.Editor && oldMode != newMode) {
-                this.enterEditorMode();
-            } else if (newMode == EcdarController.Mode.Simulator && oldMode != newMode) {
-                this.enterSimulatorMode();
-            }
-
-        });
-        if (currentMode.get() == EcdarController.Mode.Editor) {
-            this.enterEditorMode();
-        } else if (currentMode.get() == EcdarController.Mode.Simulator) {
-            this.enterSimulatorMode();
-        }
-
     }
 
     /**
@@ -266,7 +144,7 @@ public class EcdarController implements Initializable, Presentable {
                 Ecdar.getProject().getComponents().remove(newComponent);
             }, "Created new component: " + newComponent.getName(), "add-circle");
 
-            getActiveCanvasPresentation().getController().setActiveModel(newComponent);
+            MainController.getActiveCanvasPresentation().getController().setActiveModel(newComponent);
         });
         KeyboardTracker.registerKeybind(KeyboardTracker.CREATE_COMPONENT, binding);
 
@@ -445,768 +323,6 @@ public class EcdarController implements Initializable, Presentable {
         }).start();
     }
 
-    private void initializeStatusBar() {
-        statusBar.setBackground(new Background(new BackgroundFill(
-                Color.GREY_BLUE.getColor(Color.Intensity.I800),
-                CornerRadii.EMPTY,
-                Insets.EMPTY
-        )));
-
-        statusLabel.setTextFill(Color.GREY_BLUE.getColor(Color.Intensity.I50));
-        statusLabel.textProperty().bind(Ecdar.projectDirectory);
-        statusLabel.setOpacity(0.5);
-
-        queryLabel.setTextFill(Color.GREY_BLUE.getColor(Color.Intensity.I50));
-        queryLabel.setOpacity(0.5);
-
-        Debug.backgroundThreads.addListener(new ListChangeListener<Thread>() {
-            @Override
-            public void onChanged(final Change<? extends Thread> c) {
-                while (c.next()) {
-                    Platform.runLater(() -> {
-                        if(Debug.backgroundThreads.size() == 0) {
-                            queryStatusContainer.setOpacity(0);
-                        } else {
-                            queryStatusContainer.setOpacity(1);
-                            queryLabel.setText(Debug.backgroundThreads.size() + " background queries running");
-                        }
-                    });
-                }
-            }
-        });
-    }
-
-    private void initializeMenuBar() {
-        menuBar.setUseSystemMenuBar(true);
-
-        initializeCreateNewProjectMenuItem();
-        initializeOpenProjectMenuItem();
-
-        menuBarFileSave.setAccelerator(new KeyCodeCombination(KeyCode.S, KeyCombination.SHORTCUT_DOWN));
-        menuBarFileSave.setOnAction(event -> save());
-
-        menuBarFileSaveAs.setAccelerator(new KeyCodeCombination(KeyCode.S, KeyCombination.SHORTCUT_DOWN, KeyCombination.SHIFT_DOWN));
-        menuBarFileSaveAs.setOnAction(event -> saveAs());
-
-        initializeNewMutationTestObjectMenuItem();
-
-        initializeFileExportAsPng();
-
-        initializeEditMenu();
-
-        initializeViewMenu();
-
-        initializeUICacheMenuElement();
-
-        initializeHelpMenu();
-    }
-
-    public static CanvasPresentation getActiveCanvasPresentation() {
-        return activeCanvasPresentation.get();
-    }
-
-    public static void setActiveCanvasPresentation(CanvasPresentation newActiveCanvasPresentation) {
-        activeCanvasPresentation.set(newActiveCanvasPresentation);
-    }
-
-    private void initializeHelpMenu() {
-        menuBarHelpHelp.setOnAction(event -> Ecdar.showHelp());
-
-        menuBarHelpTest.setOnAction(event -> {
-            testHelpContainer.setVisible(true);
-            testHelpDialog.show(testHelpContainer);
-        });
-        testHelpAcceptButton.setOnAction(event -> testHelpDialog.close());
-        testHelpDialog.setOnDialogClosed(event -> testHelpContainer.setVisible(false)); // hide container when dialog is fully closed
-
-        menuBarHelpAbout.setOnAction(event -> {
-            aboutContainer.setVisible(true);
-            aboutDialog.show(aboutContainer);
-        });
-        aboutAcceptButton.setOnAction(event -> aboutDialog.close());
-        aboutDialog.setOnDialogClosed(event -> aboutContainer.setVisible(false)); // hide container when dialog is fully closed
-
-    }
-
-    /**
-     * Initializes the UI Cache menu element.
-     */
-    private void initializeUICacheMenuElement() {
-        menuBarOptionsCache.setOnAction(event -> {
-            final BooleanProperty isCached = Ecdar.toggleUICache();
-            menuBarOptionsCache.getGraphic().opacityProperty().bind(new When(isCached).then(1).otherwise(0));
-        });
-    }
-
-    private void initializeEditMenu() {
-        menuEditMoveLeft.setAccelerator(new KeyCodeCombination(KeyCode.LEFT, KeyCombination.CONTROL_DOWN));
-        menuEditMoveLeft.setOnAction(event -> {
-            final HighLevelModelObject activeModel = getActiveCanvasPresentation().getController().getActiveModel();
-
-            if (activeModel instanceof Component) ((Component) activeModel).moveAllNodesLeft();
-            else Ecdar.showToast("This can only be performed on components.");
-        });
-
-        menuEditMoveRight.setAccelerator(new KeyCodeCombination(KeyCode.RIGHT, KeyCombination.CONTROL_DOWN));
-        menuEditMoveRight.setOnAction(event -> {
-            final HighLevelModelObject activeModel = getActiveCanvasPresentation().getController().getActiveModel();
-
-            if (activeModel instanceof Component) ((Component) activeModel).moveAllNodesRight();
-            else Ecdar.showToast("This can only be performed on components.");
-        });
-
-        menuEditMoveUp.setAccelerator(new KeyCodeCombination(KeyCode.UP, KeyCombination.CONTROL_DOWN));
-        menuEditMoveUp.setOnAction(event -> {
-            final HighLevelModelObject activeModel = getActiveCanvasPresentation().getController().getActiveModel();
-
-            if (activeModel instanceof Component) ((Component) activeModel).moveAllNodesUp();
-            else Ecdar.showToast("This can only be performed on components.");
-        });
-
-        menuEditMoveDown.setAccelerator(new KeyCodeCombination(KeyCode.DOWN, KeyCombination.CONTROL_DOWN));
-        menuEditMoveDown.setOnAction(event -> {
-            final HighLevelModelObject activeModel = getActiveCanvasPresentation().getController().getActiveModel();
-
-            if (activeModel instanceof Component) ((Component) activeModel).moveAllNodesDown();
-            else Ecdar.showToast("This can only be performed on components.");
-        });
-    }
-
-    /**
-     * Initialize the View menu.
-     */
-    private void initializeViewMenu() {
-        menuBarViewFilePanel.getGraphic().setOpacity(1);
-        menuBarViewFilePanel.setAccelerator(new KeyCodeCombination(KeyCode.P, KeyCodeCombination.SHORTCUT_DOWN));
-        menuBarViewFilePanel.setOnAction(event -> {
-            final BooleanProperty isOpen = Ecdar.toggleFilePane();
-            menuBarViewFilePanel.getGraphic().opacityProperty().bind(new When(isOpen).then(1).otherwise(0));
-        });
-
-        menuBarViewQueryPanel.getGraphic().setOpacity(0);
-        menuBarViewQueryPanel.setAccelerator(new KeyCodeCombination(KeyCode.G, KeyCodeCombination.SHORTCUT_DOWN));
-        menuBarViewQueryPanel.setOnAction(event -> {
-            final BooleanProperty isOpen = Ecdar.toggleQueryPane();
-            menuBarViewQueryPanel.getGraphic().opacityProperty().bind(new When(isOpen).then(1).otherwise(0));
-        });
-
-        menuBarViewGrid.getGraphic().setOpacity(1);
-        menuBarViewGrid.setAccelerator(new KeyCodeCombination(KeyCode.K, KeyCodeCombination.SHORTCUT_DOWN));
-        menuBarViewGrid.setOnAction(event -> {
-            final BooleanProperty isOn = Ecdar.toggleGrid();
-            menuBarViewGrid.getGraphic().opacityProperty().bind(new When(isOn).then(1).otherwise(0));
-        });
-
-        menuBarViewCanvasSplit.getGraphic().setOpacity(1);
-        menuBarViewCanvasSplit.setOnAction(event -> {
-            final BooleanProperty isSplit = Ecdar.toggleCanvasSplit();
-            if(isSplit.get()) {
-                setCanvasModeToSingular();
-                menuBarViewCanvasSplit.setText("Split canvas");
-            } else {
-                setCanvasModeToSplit();
-                menuBarViewCanvasSplit.setText("Merge canvases");
-            }
-        });
-    }
-
-    /**
-     * Initializes the open project menu item.
-     */
-    private void initializeOpenProjectMenuItem() {
-        menuBarFileOpenProject.setAccelerator(new KeyCodeCombination(KeyCode.O, KeyCombination.SHORTCUT_DOWN));
-        menuBarFileOpenProject.setOnAction(event -> {
-            // Dialog title
-            final DirectoryChooser projectPicker = new DirectoryChooser();
-            projectPicker.setTitle("Open project");
-
-            // The initial location for the file choosing dialog
-            final File jarDir = new File(System.getProperty("java.class.path")).getAbsoluteFile().getParentFile();
-
-            // If the file does not exist, we must be running it from a development environment, use an default location
-            if(jarDir.exists()) {
-                projectPicker.setInitialDirectory(jarDir);
-            }
-
-            // Prompt the user to find a file (will halt the UI thread)
-            final File file = projectPicker.showDialog(root.getScene().getWindow());
-            if(file != null) {
-                try {
-                    Ecdar.projectDirectory.set(file.getAbsolutePath());
-                    Ecdar.initializeProjectFolder();
-                    UndoRedoStack.clear();
-                } catch (final IOException e) {
-                    e.printStackTrace();
-                }
-            }
-        });
-    }
-
-    /**
-     * Saves the project to the {@link Ecdar#projectDirectory} path.
-     * This include making directories, converting project files (components and queries)
-     * into Json formatted files.
-     */
-    public void save() {
-        if (Ecdar.projectDirectory.isNull().get()) {
-            saveAs();
-        } else {
-            save(new File(Ecdar.projectDirectory.get()));
-        }
-    }
-
-    /**
-     * Save project as.
-     */
-    private void saveAs() {
-        final FileChooser filePicker = new FileChooser();
-        filePicker.setTitle("Save project");
-        filePicker.setInitialFileName("New Ecdar Project");
-        filePicker.setInitialDirectory(new File(System.getProperty("user.home")));
-
-        final File file = filePicker.showSaveDialog(root.getScene().getWindow());
-        if (file != null) {
-            Ecdar.projectDirectory.setValue(file.getPath());
-            save(file);
-        } else {
-            Ecdar.showToast("The project was not saved.");
-        }
-    }
-
-    /**
-     * Save project at a given directory.
-     * @param directory directory to save at
-     */
-    private static void save(final File directory) {
-        try {
-            Ecdar.getProject().serialize(directory);
-        } catch (final IOException e) {
-            Ecdar.showToast("Could not save project: " + e.getMessage());
-            e.printStackTrace();
-        }
-    }
-
-    /**
-     * Initializes menu item for creating a new project.
-     */
-    private void initializeCreateNewProjectMenuItem() {
-        menuBarFileCreateNewProject.setAccelerator(new KeyCodeCombination(KeyCode.N, KeyCombination.SHORTCUT_DOWN, KeyCombination.SHIFT_DOWN));
-        menuBarFileCreateNewProject.setOnAction(event -> {
-
-            final ButtonType yesButton = new ButtonType("Save", ButtonBar.ButtonData.OK_DONE);
-            final ButtonType noButton = new ButtonType("Don't save", ButtonBar.ButtonData.NO);
-            final ButtonType cancelButton = new ButtonType("Cancel", ButtonBar.ButtonData.CANCEL_CLOSE);
-            final Alert alert = new Alert(Alert.AlertType.NONE,
-                    "Do you want to save the existing project?",
-                    yesButton,
-                    noButton,
-                    cancelButton);
-
-            alert.setTitle("Create new project");
-            final Optional<ButtonType> result = alert.showAndWait();
-
-            if (result.isPresent()) {
-                if (result.get() == yesButton) {
-                    save();
-                    createNewProject();
-                } else if (result.get() == noButton) {
-                    createNewProject();
-                }
-            }
-        });
-    }
-
-    /**
-     * Initializes menu item for creating a new mutation test object.
-     */
-    private void initializeNewMutationTestObjectMenuItem() {
-        menuBarFileNewMutationTestObject.setAccelerator(new KeyCodeCombination(KeyCode.T, KeyCombination.SHORTCUT_DOWN));
-        menuBarFileNewMutationTestObject.setOnAction(event -> {
-            final MutationTestPlan newPlan = new MutationTestPlan();
-
-            UndoRedoStack.pushAndPerform(() -> { // Perform
-                Ecdar.getProject().getTestPlans().add(newPlan);
-                getActiveCanvasPresentation().getController().setActiveModel(newPlan);
-            }, () -> { // Undo
-                Ecdar.getProject().getTestPlans().remove(newPlan);
-            }, "Created new mutation test plan", "");
-        });
-    }
-
-    /**
-     * Creates a new project.
-     */
-    private static void createNewProject() {
-        CodeAnalysis.disable();
-
-        CodeAnalysis.clearErrorsAndWarnings();
-
-        Ecdar.projectDirectory.set(null);
-
-        Ecdar.getProject().reset();
-        getActiveCanvasPresentation().getController().setActiveModel(Ecdar.getProject().getComponents().get(0));
-
-        UndoRedoStack.clear();
-
-        CodeAnalysis.enable();
-    }
-
-    /**
-     * Initializes button for exporting as png.
-     */
-    private void initializeFileExportAsPng() {
-        menuBarFileExportAsPng.setAccelerator(new KeyCodeCombination(KeyCode.L, KeyCombination.SHORTCUT_DOWN));
-        menuBarFileExportAsPng.setOnAction(event -> {
-            // If there is no active component or system
-            if (!(getActiveCanvasPresentation().getController().getActiveModel() instanceof Component || getActiveCanvasPresentation().getController().getActiveModel() instanceof EcdarSystem)) {
-                Ecdar.showToast("No component or system to export.");
-                return;
-            }
-
-            CanvasPresentation canvas = getActiveCanvasPresentation();
-
-            // If there was an active component, hide button for toggling declarations
-            final ComponentPresentation presentation = canvas.getController().getActiveComponentPresentation();
-            if (presentation != null) {
-                presentation.getController().toggleDeclarationButton.setVisible(false);
-            }
-
-            final WritableImage image = takeSnapshot(canvas);
-
-            // If there was an active component, show the button again
-            if (presentation != null) {
-                presentation.getController().toggleDeclarationButton.setVisible(true);
-            }
-
-            CropAndExportImage(image);
-        });
-
-        menuBarFileExportAsPngNoBorder.setOnAction(event -> {
-            CanvasPresentation canvas = getActiveCanvasPresentation();
-
-            final ComponentPresentation presentation = canvas.getController().getActiveComponentPresentation();
-
-            //If there is no active component
-            if (presentation == null){
-                Ecdar.showToast("No component to export.");
-                return;
-            }
-
-            presentation.getController().hideBorderAndBackground();
-            final WritableImage image = takeSnapshot(canvas);
-            presentation.getController().showBorderAndBackground();
-
-            CropAndExportImage(image);
-        });
-    }
-
-    private void initializeCanvasPane() {
-        Platform.runLater(this::setCanvasModeToSingular);
-    }
-
-    /**
-     * Removes the canvases and adds a new one, with the active component of the active canvasPresentation.
-     */
-    private void setCanvasModeToSingular() {
-        canvasPane.getChildren().clear();
-
-        CanvasShellPresentation canvasShellPresentation = new CanvasShellPresentation();
-        HighLevelModelObject model = activeCanvasPresentation.get().getController().getActiveModel();
-        if(model != null) {
-            canvasShellPresentation.getController().canvasPresentation.getController().setActiveModel(activeCanvasPresentation.get().getController().getActiveModel());
-        } else {
-            canvasShellPresentation.getController().canvasPresentation.getController().setActiveModel(Ecdar.getProject().getComponents().get(0));
-        }
-
-        canvasShellPresentation.getController().toolbar.setTranslateY(48);
-        canvasPane.getChildren().add(canvasShellPresentation);
-
-        activeCanvasPresentation.set(canvasShellPresentation.getController().canvasPresentation);
-
-        filePane.getController().updateColorsOnFilePresentations();
-    }
-
-    /**
-     * Removes the canvas and adds a GridPane with four new canvases, with different active components,
-     * the first being the one previously displayed on the single canvas.
-     */
-    private void setCanvasModeToSplit() {
-        canvasPane.getChildren().clear();
-
-        GridPane canvasGrid = new GridPane();
-
-        canvasGrid.addColumn(0);
-        canvasGrid.addColumn(1);
-        canvasGrid.addRow(0);
-        canvasGrid.addRow(1);
-
-        ColumnConstraints col1 = new ColumnConstraints();
-        col1.setPercentWidth(50);
-
-        RowConstraints row1 = new RowConstraints();
-        row1.setPercentHeight(50);
-
-        canvasGrid.getColumnConstraints().add(col1);
-        canvasGrid.getColumnConstraints().add(col1);
-        canvasGrid.getRowConstraints().add(row1);
-        canvasGrid.getRowConstraints().add(row1);
-
-        ObservableList<Component> components = Ecdar.getProject().getComponents();
-        int currentCompNum = 0, numComponents = components.size();
-
-        // Add the canvasShellPresentation at the top-left
-        CanvasShellPresentation canvasShellPresentation = initializeNewCanvasShellPresentation();
-        canvasShellPresentation.getController().canvasPresentation.getController().setActiveModel(getActiveCanvasPresentation().getController().getActiveModel());
-        canvasShellPresentation.getController().toolbar.setTranslateY(48);
-        canvasGrid.add(canvasShellPresentation, 0, 0);
-        setActiveCanvasPresentation(canvasShellPresentation.getController().canvasPresentation);
-
-        // Add the canvasShellPresentation at the top-right
-        canvasShellPresentation = initializeNewCanvasShellPresentationWithActiveComponent(components, currentCompNum);
-        canvasShellPresentation.getController().toolbar.setTranslateY(48);
-        canvasShellPresentation.setOpacity(0.75);
-        canvasGrid.add(canvasShellPresentation, 1, 0);
-
-        // Update the startIndex for the next canvasShellPresentation
-        for (int i = 0; i < numComponents; i++) {
-            if (canvasShellPresentation.getController().canvasPresentation.getController().getActiveModel() != null && canvasShellPresentation.getController().canvasPresentation.getController().getActiveModel().equals(components.get(i))) {
-                currentCompNum = i + 1;
-            }
-        }
-
-        // Add the canvasShellPresentation at the bottom-left
-        canvasShellPresentation = initializeNewCanvasShellPresentationWithActiveComponent(components, currentCompNum);
-        canvasShellPresentation.setOpacity(0.75);
-        canvasGrid.add(canvasShellPresentation, 0, 1);
-
-        // Update the startIndex for the next canvasShellPresentation
-        for (int i = 0; i < numComponents; i++)
-            if (canvasShellPresentation.getController().canvasPresentation.getController().getActiveModel() != null && canvasShellPresentation.getController().canvasPresentation.getController().getActiveModel().equals(components.get(i))) {
-                currentCompNum = i + 1;
-            }
-
-        // Add the canvasShellPresentation at the bottom-right
-        canvasShellPresentation = initializeNewCanvasShellPresentationWithActiveComponent(components, currentCompNum);
-        canvasShellPresentation.setOpacity(0.75);
-        canvasGrid.add(canvasShellPresentation, 1, 1);
-
-        canvasPane.getChildren().add(canvasGrid);
-
-        filePane.getController().updateColorsOnFilePresentations();
-    }
-
-    /**
-     * Initialize a new CanvasShellPresentation and set its active component to the next component encountered from the startIndex and return it
-     * @param components the list of components for assigning active component of the CanvasPresentation
-     * @param startIndex the index to start at when trying to find the component to set as active
-     * @return new CanvasShellPresentation
-     */
-    private CanvasShellPresentation initializeNewCanvasShellPresentationWithActiveComponent(ObservableList<Component> components, int startIndex) {
-        CanvasShellPresentation canvasShellPresentation = initializeNewCanvasShellPresentation();
-
-        int numComponents = components.size();
-        canvasShellPresentation.getController().canvasPresentation.getController().setActiveModel(null);
-        for(int currentCompNum = startIndex; currentCompNum < numComponents; currentCompNum++){
-            if(getActiveCanvasPresentation().getController().getActiveModel() != components.get(currentCompNum)) {
-                canvasShellPresentation.getController().canvasPresentation.getController().setActiveModel(components.get(currentCompNum));
-                break;
-            }
-        }
-
-        return canvasShellPresentation;
-    }
-
-    /**
-     * Initialize a new CanvasShellPresentation and return it
-     * @return new CanvasShellPresentation
-     */
-    private CanvasShellPresentation initializeNewCanvasShellPresentation() {
-        CanvasShellPresentation canvasShellPresentation = new CanvasShellPresentation();
-        canvasShellPresentation.setBorder(new Border(new BorderStroke(Color.GREY.getColor(Color.Intensity.I500), BorderStrokeStyle.SOLID, CornerRadii.EMPTY, BorderStroke.THIN)));
-        return canvasShellPresentation;
-    }
-
-    /**
-     * Take a snapshot with the grid hidden.
-     * The grid is put into its original state afterwards.
-     * @return the snapshot
-     */
-    private WritableImage takeSnapshot(CanvasPresentation canvas) {
-        final WritableImage image;
-
-        canvas.getController().disallowGrid();
-        image = scaleAndTakeSnapshot(canvas);
-        canvas.getController().allowGrid();
-
-        return image;
-    }
-
-    /**
-     * Zooms in times 4 to get a higher resolution.
-     * Then take snapshot and zoom to times 1 again.
-     * @return the snapshot
-     */
-    private WritableImage scaleAndTakeSnapshot(CanvasPresentation canvas) {
-        final WritableImage image;
-        Double zoomLevel = canvas.getController().zoomHelper.getZoomLevel();
-        canvas.getController().zoomHelper.zoomToFit();
-
-        image = canvas.snapshot(new SnapshotParameters(), null);
-
-        canvas.getController().zoomHelper.setZoomLevel(zoomLevel);
-        return image;
-    }
-
-    /**
-     * Crops and exports an image.
-     * @param image the image
-     */
-    private void CropAndExportImage(final WritableImage image) {
-        final String name = getActiveCanvasPresentation().getController().getActiveModel().getName();
-
-        final FileChooser filePicker = new FileChooser();
-        filePicker.setTitle("Export png");
-        filePicker.setInitialFileName(name + ".png");
-
-        // Set initial directory to project directory (if saved) or user.home (otherwise)
-        String directory = Ecdar.projectDirectory.get();
-        if (directory == null) directory = System.getProperty("user.home");
-
-        filePicker.setInitialDirectory(new File(directory));
-        filePicker.getExtensionFilters().add(new FileChooser.ExtensionFilter("PNG File", "*.png"));
-
-        final BufferedImage finalImage;
-        try {
-            finalImage = autoCropImage(SwingFXUtils.fromFXImage(image, null));
-        } catch (final IllegalArgumentException e) {
-            Ecdar.showToast("Export failed. " + e.getMessage());
-            return;
-        }
-
-        final File file = filePicker.showSaveDialog(root.getScene().getWindow());
-        if (file != null){
-            try {
-                ImageIO.write(finalImage, "png", file);
-                Ecdar.showToast("Export succeeded.");
-            } catch (final IOException e) {
-                Ecdar.showToast("Export failed. "+ e.getMessage());
-            }
-        } else {
-            Ecdar.showToast("Export was cancelled.");
-        }
-    }
-
-    /**
-     * Crops an image so that the all-white borders are removed.
-     * @param image the original image
-     * @return the cropped image
-     */
-    private static BufferedImage autoCropImage(final BufferedImage image) {
-        final int topY = getAutoCropTopY(image);
-        final int leftX = getAutoCropLeftX(image);
-        final int rightX = getAutoCropRightX(image);
-        final int bottomY = getAutoCropBottomY(image);
-
-
-        return image.getSubimage(leftX, topY, rightX - leftX, bottomY - topY);
-    }
-
-    /**
-     * Gets the top y coordinate of an auto cropped image.
-     * @param image the original image
-     * @return the y coordinate
-     */
-    private static int getAutoCropTopY(final BufferedImage image) {
-        for (int y = 0; y < image.getHeight(); y++) {
-            for (int x = 0; x < image.getWidth(); x++) {
-                if (image.getRGB(x, y) != java.awt.Color.WHITE.getRGB()) {
-                    return y;
-                }
-            }
-        }
-
-        throw new IllegalArgumentException("Image is all white");
-    }
-
-    /**
-     * Gets the left x coordinate of an auto cropped image.
-     * @param image the original image
-     * @return the x coordinate
-     */
-    private static int getAutoCropLeftX(final BufferedImage image) {
-        for (int x = 0; x < image.getWidth(); x++) {
-            for (int y = 0; y < image.getHeight(); y++) {
-                if (image.getRGB(x, y) != java.awt.Color.WHITE.getRGB()) {
-                    return x;
-                }
-            }
-        }
-
-        throw new IllegalArgumentException("Image is all white");
-    }
-
-    /**
-     * Gets the bottom y coordinate of an auto cropped image.
-     * @param image the original image
-     * @return the y coordinate
-     */
-    private static int getAutoCropBottomY(final BufferedImage image) {
-        for (int y = image.getHeight() - 1; y >= 0; y--) {
-            for (int x = 0; x < image.getWidth(); x++) {
-                if (image.getRGB(x, y) != java.awt.Color.WHITE.getRGB()) {
-                    return y;
-                }
-            }
-        }
-
-        throw new IllegalArgumentException("Image is all white");
-    }
-
-    /**
-     * Gets the right x coordinate of an auto cropped image.
-     * @param image the original image
-     * @return the x coordinate
-     */
-    private static int getAutoCropRightX(final BufferedImage image) {
-        for (int x = image.getWidth() - 1; x >= 0; x--) {
-            for (int y = 0; y < image.getHeight(); y++) {
-                if (image.getRGB(x, y) != java.awt.Color.WHITE.getRGB()) {
-                    return x;
-                }
-            }
-        }
-
-        throw new IllegalArgumentException("Image is all white");
-    }
-
-    private void initializeMessages() {
-        final Map<Component, MessageCollectionPresentation> componentMessageCollectionPresentationMapForErrors = new HashMap<>();
-        final Map<Component, MessageCollectionPresentation> componentMessageCollectionPresentationMapForWarnings = new HashMap<>();
-
-        final Consumer<Component> addComponent = (component) -> {
-            final MessageCollectionPresentation messageCollectionPresentationErrors = new MessageCollectionPresentation(component, CodeAnalysis.getErrors(component));
-            componentMessageCollectionPresentationMapForErrors.put(component, messageCollectionPresentationErrors);
-            errorsList.getChildren().add(messageCollectionPresentationErrors);
-
-            final Runnable addIfErrors = () -> {
-                if (CodeAnalysis.getErrors(component).size() == 0) {
-                    errorsList.getChildren().remove(messageCollectionPresentationErrors);
-                } else if (!errorsList.getChildren().contains(messageCollectionPresentationErrors)) {
-                    errorsList.getChildren().add(messageCollectionPresentationErrors);
-                }
-            };
-
-            addIfErrors.run();
-            CodeAnalysis.getErrors(component).addListener(new ListChangeListener<CodeAnalysis.Message>() {
-                @Override
-                public void onChanged(final Change<? extends CodeAnalysis.Message> c) {
-                    while (c.next()) {
-                        addIfErrors.run();
-                    }
-                }
-            });
-
-            final MessageCollectionPresentation messageCollectionPresentationWarnings = new MessageCollectionPresentation(component, CodeAnalysis.getWarnings(component));
-            componentMessageCollectionPresentationMapForWarnings.put(component, messageCollectionPresentationWarnings);
-            warningsList.getChildren().add(messageCollectionPresentationWarnings);
-
-            final Runnable addIfWarnings = () -> {
-                if (CodeAnalysis.getWarnings(component).size() == 0) {
-                    warningsList.getChildren().remove(messageCollectionPresentationWarnings);
-                } else if (!warningsList.getChildren().contains(messageCollectionPresentationWarnings)) {
-                    warningsList.getChildren().add(messageCollectionPresentationWarnings);
-                }
-            };
-
-            addIfWarnings.run();
-            CodeAnalysis.getWarnings(component).addListener(new ListChangeListener<CodeAnalysis.Message>() {
-                @Override
-                public void onChanged(final Change<? extends CodeAnalysis.Message> c) {
-                    while (c.next()) {
-                        addIfWarnings.run();
-                    }
-                }
-            });
-        };
-
-        // Add error that is project wide but not a backend error
-        addComponent.accept(null);
-
-        Ecdar.getProject().getComponents().forEach(addComponent);
-        Ecdar.getProject().getComponents().addListener(new ListChangeListener<Component>() {
-            @Override
-            public void onChanged(final Change<? extends Component> c) {
-                while (c.next()) {
-                    c.getAddedSubList().forEach(addComponent::accept);
-
-                    c.getRemoved().forEach(component -> {
-                        errorsList.getChildren().remove(componentMessageCollectionPresentationMapForErrors.get(component));
-                        componentMessageCollectionPresentationMapForErrors.remove(component);
-
-                        warningsList.getChildren().remove(componentMessageCollectionPresentationMapForWarnings.get(component));
-                        componentMessageCollectionPresentationMapForWarnings.remove(component);
-                    });
-                }
-            }
-        });
-
-        final Map<CodeAnalysis.Message, MessagePresentation> messageMessagePresentationHashMap = new HashMap<>();
-
-        CodeAnalysis.getBackendErrors().addListener(new ListChangeListener<CodeAnalysis.Message>() {
-            @Override
-            public void onChanged(final Change<? extends CodeAnalysis.Message> c) {
-                while (c.next()) {
-                    c.getAddedSubList().forEach(addedMessage -> {
-                        final MessagePresentation messagePresentation = new MessagePresentation(addedMessage);
-                        backendErrorsList.getChildren().add(messagePresentation);
-                        messageMessagePresentationHashMap.put(addedMessage, messagePresentation);
-                    });
-
-                    c.getRemoved().forEach(removedMessage -> {
-                        backendErrorsList.getChildren().remove(messageMessagePresentationHashMap.get(removedMessage));
-                        messageMessagePresentationHashMap.remove(removedMessage);
-                    });
-                }
-            }
-        });
-    }
-
-    private void initializeTabPane() {
-        bottomFillerElement.heightProperty().bind(tabPaneContainer.maxHeightProperty());
-
-        tabPane.getSelectionModel().selectedIndexProperty().addListener((obs, oldSelected, newSelected) -> {
-            if (newSelected.intValue() < 0 || tabPaneContainer.getMaxHeight() > 35) return;
-
-            if (shouldISkipOpeningTheMessagesContainer) {
-                tabPane.getSelectionModel().clearSelection();
-                shouldISkipOpeningTheMessagesContainer = false;
-            } else {
-                expandMessagesIfNotExpanded();
-            }
-        });
-
-        tabPane.getSelectionModel().clearSelection();
-
-        tabPane.setTabMinHeight(35);
-        tabPane.setTabMaxHeight(35);
-    }
-
-    @FXML
-    private void tabPaneResizeElementPressed(final MouseEvent event) {
-        tabPanePreviousY = event.getScreenY();
-    }
-
-    @FXML
-    private void tabPaneResizeElementDragged(final MouseEvent event) {
-        final double mouseY = event.getScreenY();
-        double newHeight = tabPaneContainer.getMaxHeight() - (mouseY - tabPanePreviousY);
-        newHeight = Math.max(35, newHeight);
-
-        setMaxHeight(newHeight);
-        tabPanePreviousY = mouseY;
-
-    }
-
     public void expandMessagesIfNotExpanded() {
         if (tabPaneContainer.getMaxHeight() <= 35) {
             expandMessagesContainer.play();
@@ -1256,77 +372,6 @@ public class EcdarController implements Initializable, Presentable {
         } else {
             expandMessagesContainer.play();
         }
-    }
-
-    private void enterEditorMode() {
-        if (!this.centerContainer.getChildren().contains(Ecdar.getPresentation())) {
-            Ecdar.getPresentation().getController().willShow();
-            this.simulatorPresentation.getController().willHide();
-            this.centerContainer.getChildren().remove(this.simulatorPresentation);
-            this.centerContainer.getChildren().add(Ecdar.getPresentation());
-            this.updateMenuItems();
-        }
-    }
-
-    private void enterSimulatorMode() {
-        if (!this.centerContainer.getChildren().contains(this.simulatorPresentation)) {
-            currentMode.setValue(EcdarController.Mode.Simulator);
-            Ecdar.getPresentation().getController().willHide();
-            this.simulatorPresentation.getController().willShow();
-            this.centerContainer.getChildren().remove(Ecdar.getPresentation());
-            this.centerContainer.getChildren().add(this.simulatorPresentation);
-            this.updateMenuItems();
-        }
-    }
-
-    private void updateMenuItems() {
-        switch(currentMode.get()) {
-            case Editor:
-                this.menuBarViewGrid.setDisable(false);
-                this.menuBarViewFilePanel.setDisable(false);
-                this.menuBarViewQueryPanel.setDisable(false);
-                this.menuBarViewSimLeftPanel.setDisable(true);
-                this.menuBarViewSimRightPanel.setDisable(true);
-                this.menuBarFileExportAsPngNoBorder.setDisable(false);
-                this.menuBarFileExportAsPng.setDisable(false);
-                this.menuBarZoomInSimulator.setDisable(true);
-                this.menuBarZoomOutSimulator.setDisable(true);
-                this.menuBarZoomResetSimulator.setDisable(true);
-                KeyboardTracker.unregisterKeybind("ZOOM_IN");
-                KeyboardTracker.unregisterKeybind("ZOOM_OUT");
-                KeyboardTracker.unregisterKeybind("ZOOM_RESET");
-                break;
-            case Simulator:
-                this.menuBarViewGrid.setDisable(true);
-                this.menuBarViewFilePanel.setDisable(true);
-                this.menuBarViewSimLeftPanel.setDisable(false);
-                this.menuBarViewSimRightPanel.setDisable(false);
-                this.menuBarFileExportAsPngNoBorder.setDisable(true);
-                this.menuBarFileExportAsPng.setDisable(true);
-                this.menuBarViewQueryPanel.setDisable(true);
-                this.menuBarZoomInSimulator.setDisable(false);
-                this.menuBarZoomOutSimulator.setDisable(false);
-                this.menuBarZoomResetSimulator.setDisable(false);
-                this.setExtraZoomKeybindings();
-        }
-    }
-
-    private void setExtraZoomKeybindings() {
-        KeyCodeCombination zoomInCombination2 = new KeyCodeCombination(KeyCode.ADD, KeyCombination.SHORTCUT_DOWN);
-        KeyboardTracker.registerKeybind("ZOOM_IN", new Keybind(zoomInCombination2, (keyEvent) -> {
-            keyEvent.consume();
-            this.simulatorPresentation.getController().overviewPresentation.getController().zoomIn();
-        }));
-        KeyCodeCombination zoomOutCombination2 = new KeyCodeCombination(KeyCode.SUBTRACT, KeyCombination.SHORTCUT_DOWN);
-        KeyboardTracker.registerKeybind("ZOOM_OUT", new Keybind(zoomOutCombination2, (keyEvent) -> {
-            keyEvent.consume();
-            this.simulatorPresentation.getController().overviewPresentation.getController().zoomOut();
-        }));
-        KeyCodeCombination zoomResetCombination = new KeyCodeCombination(KeyCode.NUMPAD0, KeyCombination.SHORTCUT_DOWN);
-        KeyboardTracker.registerKeybind("ZOOM_RESET", new Keybind(zoomResetCombination, (keyEvent) -> {
-            keyEvent.consume();
-            this.simulatorPresentation.getController().overviewPresentation.getController().resetZoom();
-        }));
     }
 
     /**
@@ -1456,55 +501,9 @@ public class EcdarController implements Initializable, Presentable {
         globalEdgeStatus.set(status);
     }
 
-    @FXML
-    private void closeDialog() {
-        dialog.close();
-        queryDialog.close();
-    }
-
-    public static void openQueryDialog(final Query query, final String text) {
-        if (text != null) {
-            _queryTextResult.setText(text);
-        }
-        if (query != null) {
-            _queryTextQuery.setText(query.getQuery());
-        }
-        _queryDialog.show();
-    }
-
-    /**
-     * Used by the cancel button in the reload dialog.
-     * Closes the dialog when called.
-     */
-    @FXML
-    private void cancelReloadSimulatorDialog() {
-        reloadSimulatorDialog.close();
-    }
-
-    /**
-     * Used by the OK button in the reload dialog.
-     * Reloads the system of the simulator.
-     */
-    @FXML
-    private void okReloadSimulatorDialog() {
-        simulatorPresentation.getController().resetCurrentSimulation();
-        reloadSimulatorDialog.close();
-    }
-
-    public static void openReloadSimulationDialog() {
-        _reloadSimulatorDialog.show();
-    }
-
-    /**
-     * Changes the mode to simulator, causing the view to change to simulator mode
-     */
-    public static void showSimulator() {
-        currentMode.setValue(Mode.Simulator);
-    }
-
     public void willShow() {
         initializeKeybindings();
-        getActiveCanvasPresentation().getController().resetCurrentActiveModelPlacement();
+        MainController.getActiveCanvasPresentation().getController().resetCurrentActiveModelPlacement();
     }
 
     public void willHide() {
@@ -1526,11 +525,137 @@ public class EcdarController implements Initializable, Presentable {
         KeyboardTracker.unregisterKeybind("COLOR_SELECTED");
     }
 
-    private static enum Mode {
-        Editor,
-        Simulator;
-
-        private Mode() {
+    public void setCanvasSplit(boolean shouldSplit) {
+        if (shouldSplit) {
+            setCanvasModeToSplit();
+        } else {
+            setCanvasModeToSingular();
         }
+    }
+
+    private void initializeCanvasPane() {
+        Platform.runLater(this::setCanvasModeToSingular);
+    }
+
+    /**
+     * Initialize a new CanvasShellPresentation and return it
+     * @return new CanvasShellPresentation
+     */
+    private CanvasShellPresentation initializeNewCanvasShellPresentation() {
+        CanvasShellPresentation canvasShellPresentation = new CanvasShellPresentation();
+        canvasShellPresentation.setBorder(new Border(new BorderStroke(Color.GREY.getColor(Color.Intensity.I500), BorderStrokeStyle.SOLID, CornerRadii.EMPTY, BorderStroke.THIN)));
+        return canvasShellPresentation;
+    }
+
+    /**
+     * Initialize a new CanvasShellPresentation and set its active component to the next component encountered from the startIndex and return it
+     * @param components the list of components for assigning active component of the CanvasPresentation
+     * @param startIndex the index to start at when trying to find the component to set as active
+     * @return new CanvasShellPresentation
+     */
+    private CanvasShellPresentation initializeNewCanvasShellPresentationWithActiveComponent(ObservableList<Component> components, int startIndex) {
+        CanvasShellPresentation canvasShellPresentation = initializeNewCanvasShellPresentation();
+
+        int numComponents = components.size();
+        canvasShellPresentation.getController().canvasPresentation.getController().setActiveModel(null);
+        for(int currentCompNum = startIndex; currentCompNum < numComponents; currentCompNum++){
+            if(MainController.getActiveCanvasPresentation().getController().getActiveModel() != components.get(currentCompNum)) {
+                canvasShellPresentation.getController().canvasPresentation.getController().setActiveModel(components.get(currentCompNum));
+                break;
+            }
+        }
+
+        return canvasShellPresentation;
+    }
+
+    /**
+     * Removes the canvases and adds a new one, with the active component of the active canvasPresentation.
+     */
+    private void setCanvasModeToSingular() {
+        canvasPane.getChildren().clear();
+
+        CanvasShellPresentation canvasShellPresentation = new CanvasShellPresentation();
+        HighLevelModelObject model = MainController.getActiveCanvasPresentation().getController().getActiveModel();
+        if(model != null) {
+            canvasShellPresentation.getController().canvasPresentation.getController().setActiveModel(MainController.getActiveCanvasPresentation().getController().getActiveModel());
+        } else {
+            canvasShellPresentation.getController().canvasPresentation.getController().setActiveModel(Ecdar.getProject().getComponents().get(0));
+        }
+
+        canvasShellPresentation.getController().toolbar.setTranslateY(48);
+        canvasPane.getChildren().add(canvasShellPresentation);
+
+        MainController.setActiveCanvasPresentation(canvasShellPresentation.getController().canvasPresentation);
+
+        filePane.getController().updateColorsOnFilePresentations();
+    }
+
+    /**
+     * Removes the canvas and adds a GridPane with four new canvases, with different active components,
+     * the first being the one previously displayed on the single canvas.
+     */
+    private void setCanvasModeToSplit() {
+        canvasPane.getChildren().clear();
+
+        GridPane canvasGrid = new GridPane();
+
+        canvasGrid.addColumn(0);
+        canvasGrid.addColumn(1);
+        canvasGrid.addRow(0);
+        canvasGrid.addRow(1);
+
+        ColumnConstraints col1 = new ColumnConstraints();
+        col1.setPercentWidth(50);
+
+        RowConstraints row1 = new RowConstraints();
+        row1.setPercentHeight(50);
+
+        canvasGrid.getColumnConstraints().add(col1);
+        canvasGrid.getColumnConstraints().add(col1);
+        canvasGrid.getRowConstraints().add(row1);
+        canvasGrid.getRowConstraints().add(row1);
+
+        ObservableList<Component> components = Ecdar.getProject().getComponents();
+        int currentCompNum = 0, numComponents = components.size();
+
+        // Add the canvasShellPresentation at the top-left
+        CanvasShellPresentation canvasShellPresentation = initializeNewCanvasShellPresentation();
+        canvasShellPresentation.getController().canvasPresentation.getController().setActiveModel(MainController.getActiveCanvasPresentation().getController().getActiveModel());
+        canvasShellPresentation.getController().toolbar.setTranslateY(48);
+        canvasGrid.add(canvasShellPresentation, 0, 0);
+        MainController.setActiveCanvasPresentation(canvasShellPresentation.getController().canvasPresentation);
+
+        // Add the canvasShellPresentation at the top-right
+        canvasShellPresentation = initializeNewCanvasShellPresentationWithActiveComponent(components, currentCompNum);
+        canvasShellPresentation.getController().toolbar.setTranslateY(48);
+        canvasShellPresentation.setOpacity(0.75);
+        canvasGrid.add(canvasShellPresentation, 1, 0);
+
+        // Update the startIndex for the next canvasShellPresentation
+        for (int i = 0; i < numComponents; i++) {
+            if (canvasShellPresentation.getController().canvasPresentation.getController().getActiveModel() != null && canvasShellPresentation.getController().canvasPresentation.getController().getActiveModel().equals(components.get(i))) {
+                currentCompNum = i + 1;
+            }
+        }
+
+        // Add the canvasShellPresentation at the bottom-left
+        canvasShellPresentation = initializeNewCanvasShellPresentationWithActiveComponent(components, currentCompNum);
+        canvasShellPresentation.setOpacity(0.75);
+        canvasGrid.add(canvasShellPresentation, 0, 1);
+
+        // Update the startIndex for the next canvasShellPresentation
+        for (int i = 0; i < numComponents; i++)
+            if (canvasShellPresentation.getController().canvasPresentation.getController().getActiveModel() != null && canvasShellPresentation.getController().canvasPresentation.getController().getActiveModel().equals(components.get(i))) {
+                currentCompNum = i + 1;
+            }
+
+        // Add the canvasShellPresentation at the bottom-right
+        canvasShellPresentation = initializeNewCanvasShellPresentationWithActiveComponent(components, currentCompNum);
+        canvasShellPresentation.setOpacity(0.75);
+        canvasGrid.add(canvasShellPresentation, 1, 1);
+
+        canvasPane.getChildren().add(canvasGrid);
+
+        filePane.getController().updateColorsOnFilePresentations();
     }
 }
