+++ src/main/java/ecdar/mutation/TextFlowBuilder.java
@@ -6,6 +6,7 @@ import ecdar.abstractions.Edge;
 import ecdar.abstractions.Location;
 import ecdar.controllers.CanvasController;
 import ecdar.controllers.EcdarController;
+import ecdar.controllers.MainController;
 import ecdar.utility.helpers.SelectHelper;
 import javafx.application.Platform;
 import javafx.collections.FXCollections;
@@ -95,7 +96,7 @@ public class TextFlowBuilder {
             }
 
             SelectHelper.elementsToBeSelected = FXCollections.observableArrayList();
-            EcdarController.getActiveCanvasPresentation().getController().setActiveModel(component);
+            MainController.getActiveCanvasPresentation().getController().setActiveModel(component);
 
             // Use a list, since there could be multiple locations (e.i. Universal locations)
             final List<Location> locations = component.getLocations().filtered(loc -> loc.getId().equals(locationId));
