+++ src/main/java/ecdar/presentations/EcdarPresentation.java
@@ -1,32 +1,25 @@
 package ecdar.presentations;
 
-import com.jfoenix.controls.JFXSnackbarLayout;
 import ecdar.Ecdar;
 import ecdar.abstractions.Query;
-import ecdar.code_analysis.CodeAnalysis;
 import ecdar.controllers.EcdarController;
+import ecdar.controllers.MainController;
 import ecdar.utility.UndoRedoStack;
 import ecdar.utility.colors.Color;
 import ecdar.utility.colors.EnabledColor;
 import ecdar.utility.helpers.SelectHelper;
 import com.jfoenix.controls.JFXPopup;
 import com.jfoenix.controls.JFXRippler;
-import com.jfoenix.controls.JFXSnackbar;
 import ecdar.utility.keyboard.Keybind;
 import ecdar.utility.keyboard.KeyboardTracker;
 import javafx.animation.*;
-import javafx.beans.InvalidationListener;
-import javafx.beans.Observable;
 import javafx.beans.property.BooleanProperty;
 import javafx.beans.property.SimpleBooleanProperty;
 import javafx.beans.property.SimpleDoubleProperty;
 import javafx.collections.ListChangeListener;
 import javafx.geometry.Insets;
-import javafx.scene.Cursor;
-import javafx.scene.Node;
 import javafx.scene.control.Label;
 import javafx.scene.control.Tooltip;
-import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
 import javafx.scene.input.KeyCode;
 import javafx.scene.input.KeyCodeCombination;
@@ -56,9 +49,8 @@ public class EcdarPresentation extends StackPane {
 
     public EcdarPresentation() {
         controller = new EcdarFXMLLoader().loadAndGetController("EcdarPresentation.fxml", this);
-        initializeTopBar();
+
         initializeToolbar();
-        initializeQueryDetailsDialog();
         initializeColorSelector();
 
         initializeToggleQueryPaneFunctionality();
@@ -74,9 +66,6 @@ public class EcdarPresentation extends StackPane {
         initializeToolbarButton(controller.redo);
         initializeUndoRedoButtons();
 
-        initializeMessageContainer();
-        initializeSnackbar();
-
         // Open the file and query panel initially
         final BooleanProperty ranInitialToggle = new SimpleBooleanProperty(false);
         controller.filePane.widthProperty().addListener((observable) -> {
@@ -84,131 +73,13 @@ public class EcdarPresentation extends StackPane {
             toggleFilePane();
             ranInitialToggle.set(true);
         });
-        initializeHelpImages();
 
-        KeyboardTracker.registerKeybind(KeyboardTracker.ZOOM_IN, new Keybind(new KeyCodeCombination(KeyCode.PLUS, KeyCombination.SHORTCUT_DOWN), () -> EcdarController.getActiveCanvasPresentation().getController().zoomHelper.zoomIn()));
-        KeyboardTracker.registerKeybind(KeyboardTracker.ZOOM_OUT, new Keybind(new KeyCodeCombination(KeyCode.MINUS, KeyCombination.SHORTCUT_DOWN), () -> EcdarController.getActiveCanvasPresentation().getController().zoomHelper.zoomOut()));
-        KeyboardTracker.registerKeybind(KeyboardTracker.ZOOM_TO_FIT, new Keybind(new KeyCodeCombination(KeyCode.EQUALS, KeyCombination.SHORTCUT_DOWN), () -> EcdarController.getActiveCanvasPresentation().getController().zoomHelper.zoomToFit()));
-        KeyboardTracker.registerKeybind(KeyboardTracker.RESET_ZOOM, new Keybind(new KeyCodeCombination(KeyCode.DIGIT0, KeyCombination.SHORTCUT_DOWN), () -> EcdarController.getActiveCanvasPresentation().getController().zoomHelper.resetZoom()));
+        KeyboardTracker.registerKeybind(KeyboardTracker.ZOOM_IN, new Keybind(new KeyCodeCombination(KeyCode.PLUS, KeyCombination.SHORTCUT_DOWN), () -> MainController.getActiveCanvasPresentation().getController().zoomHelper.zoomIn()));
+        KeyboardTracker.registerKeybind(KeyboardTracker.ZOOM_OUT, new Keybind(new KeyCodeCombination(KeyCode.MINUS, KeyCombination.SHORTCUT_DOWN), () -> MainController.getActiveCanvasPresentation().getController().zoomHelper.zoomOut()));
+        KeyboardTracker.registerKeybind(KeyboardTracker.ZOOM_TO_FIT, new Keybind(new KeyCodeCombination(KeyCode.EQUALS, KeyCombination.SHORTCUT_DOWN), () -> MainController.getActiveCanvasPresentation().getController().zoomHelper.zoomToFit()));
+        KeyboardTracker.registerKeybind(KeyboardTracker.RESET_ZOOM, new Keybind(new KeyCodeCombination(KeyCode.DIGIT0, KeyCombination.SHORTCUT_DOWN), () -> MainController.getActiveCanvasPresentation().getController().zoomHelper.resetZoom()));
         KeyboardTracker.registerKeybind(KeyboardTracker.UNDO, new Keybind(new KeyCodeCombination(KeyCode.Z, KeyCombination.SHORTCUT_DOWN), UndoRedoStack::undo));
         KeyboardTracker.registerKeybind(KeyboardTracker.REDO, new Keybind(new KeyCodeCombination(KeyCode.Z, KeyCombination.SHORTCUT_DOWN, KeyCombination.SHIFT_DOWN), UndoRedoStack::redo));
-    }
-
-    public BooleanProperty toggleLeftSimPane() {
-        return this.controller.simulatorPresentation.toggleLeftPane();
-    }
-
-    public BooleanProperty toggleRightSimPane() {
-        return this.controller.simulatorPresentation.toggleRightPane();
-    }
-
-    private void initializeSnackbar() {
-        controller.snackbar = new JFXSnackbar(controller.root);
-        controller.snackbar.setPrefWidth(568);
-        controller.snackbar.autosize();
-    }
-
-    private void initializeMessageContainer() {
-        // The element of which you drag to resize should be equal to the width of the window (main stage)
-        controller.tabPaneResizeElement.sceneProperty().addListener((observableScene, oldScene, newScene) -> {
-            if (oldScene == null && newScene != null) {
-                // scene is set for the first time. Now its the time to listen stage changes.
-                newScene.windowProperty().addListener((observableWindow, oldWindow, newWindow) -> {
-                    if (oldWindow == null && newWindow != null) {
-                        newWindow.widthProperty().addListener((observableWidth, oldWidth, newWidth) -> {
-                            controller.tabPaneResizeElement.setWidth(newWidth.doubleValue() - 30);
-                        });
-                    }
-                });
-            }
-        });
-
-        // Resize cursor
-        controller.tabPaneResizeElement.setCursor(Cursor.N_RESIZE);
-
-        controller.tabPaneContainer.maxHeightProperty().addListener((obs, oldHeight, newHeight) -> {
-            if (newHeight.doubleValue() > 35) {
-                controller.collapseMessagesIcon.setIconLiteral("gmi-close");
-                controller.collapseMessagesIcon.setIconSize(24);
-            } else {
-                controller.tabPane.getSelectionModel().clearSelection(); // Clear the currently selected tab (so that the view will open again when selecting a tab)
-                controller.collapseMessagesIcon.setIconLiteral("gmi-expand-less");
-                controller.collapseMessagesIcon.setIconSize(24);
-            }
-        });
-
-        // Remove the background of the scroll panes
-        controller.errorsScrollPane.setStyle("-fx-background-color: transparent;");
-        controller.warningsScrollPane.setStyle("-fx-background-color: transparent;");
-
-        final Runnable collapseIfNoErrorsOrWarnings = () -> {
-            new Thread(() -> {
-                // Wait for a second to check if new warnings or errors occur
-                try {
-                    Thread.sleep(1000);
-                } catch (InterruptedException e) {
-                    e.printStackTrace();
-                }
-
-                // Check if any warnings or errors occurred
-                if (CodeAnalysis.getBackendErrors().size() + CodeAnalysis.getErrors().size() + CodeAnalysis.getWarnings().size() == 0) {
-                    controller.collapseMessagesIfNotCollapsed();
-                }
-            }).start();
-        };
-
-        // Update the tab-text and expand/collapse the view
-        CodeAnalysis.getBackendErrors().addListener(new InvalidationListener() {
-            @Override
-            public void invalidated(final Observable observable) {
-                final int errors = CodeAnalysis.getBackendErrors().size();
-                if (errors == 0) {
-                    controller.backendErrorsTab.setText("Backend Errors");
-                } else {
-                    controller.backendErrorsTab.setText("Backend Errors (" + errors + ")");
-                    controller.expandMessagesIfNotExpanded();
-                    controller.tabPane.getSelectionModel().select(controller.backendErrorsTab);
-                }
-
-                collapseIfNoErrorsOrWarnings.run();
-            }
-        });
-
-        // Update the tab-text and expand/collapse the view
-        CodeAnalysis.getErrors().addListener(new InvalidationListener() {
-            @Override
-            public void invalidated(final Observable observable) {
-                final int errors = CodeAnalysis.getErrors().size();
-                if (errors == 0) {
-                    controller.errorsTab.setText("Errors");
-                } else {
-                    controller.errorsTab.setText("Errors (" + errors + ")");
-                    controller.expandMessagesIfNotExpanded();
-                    controller.tabPane.getSelectionModel().select(controller.errorsTab);
-                }
-
-                collapseIfNoErrorsOrWarnings.run();
-            }
-        });
-
-
-        // Update the tab-text and expand/collapse the view
-        CodeAnalysis.getWarnings().addListener(new InvalidationListener() {
-            @Override
-            public void invalidated(final Observable observable) {
-                final int warnings = CodeAnalysis.getWarnings().size();
-                if (warnings == 0) {
-                    controller.warningsTab.setText("Warnings");
-                } else {
-                    controller.warningsTab.setText("Warnings (" + warnings + ")");
-                    //We must select the warnings tab but we don't want the messages areas to open
-                    controller.shouldISkipOpeningTheMessagesContainer = true;
-                    controller.tabPane.getSelectionModel().select(controller.warningsTab);
-                }
-
-                collapseIfNoErrorsOrWarnings.run();
-            }
-        });
     }
 
     private void initializeUndoRedoButtons() {
@@ -353,18 +224,6 @@ public class EcdarPresentation extends StackPane {
         button.setPosition(JFXRippler.RipplerPos.BACK);
     }
 
-    private void initializeQueryDetailsDialog() {
-        final Color modalBarColor = Color.GREY_BLUE;
-        final Color.Intensity modalBarColorIntensity = Color.Intensity.I500;
-
-        // Set the background of the modal bar
-        controller.modalBar.setBackground(new Background(new BackgroundFill(
-                modalBarColor.getColor(modalBarColorIntensity),
-                CornerRadii.EMPTY,
-                Insets.EMPTY
-        )));
-    }
-
     private void initializeToggleQueryPaneFunctionality() {
         // Set the translation of the query pane to be equal to its width
         // Will hide the element, and force it in then the right side of the border pane is enlarged
@@ -478,26 +337,6 @@ public class EcdarPresentation extends StackPane {
         closeFilePaneAnimation.getKeyFrames().addAll(kf1, kf2);
     }
 
-    private void initializeTopBar() {
-        final Color color = Color.GREY_BLUE;
-        final Color.Intensity intensity = Color.Intensity.I800;
-
-        // Set the background for the top toolbar
-        controller.menuBar.setBackground(
-                new Background(new BackgroundFill(color.getColor(intensity),
-                        CornerRadii.EMPTY,
-                        Insets.EMPTY)
-                ));
-
-        // Set the bottom border
-        controller.menuBar.setBorder(new Border(new BorderStroke(
-                color.getColor(intensity.next()),
-                BorderStrokeStyle.SOLID,
-                CornerRadii.EMPTY,
-                new BorderWidths(0, 0, 1, 0)
-        )));
-    }
-
     private void initializeToolbar() {
         final Color color = Color.GREY_BLUE;
         final Color.Intensity intensity = Color.Intensity.I700;
@@ -508,23 +347,6 @@ public class EcdarPresentation extends StackPane {
                         CornerRadii.EMPTY,
                         Insets.EMPTY)
                 ));
-    }
-
-    /**
-     * Initialize help image views.
-     */
-    private void initializeHelpImages() {
-        controller.helpInitialImage.setImage(new Image(Ecdar.class.getResource("ic_help_initial.png").toExternalForm()));
-        fitSizeWhenAvailable(controller.helpInitialImage, controller.helpInitialPane);
-
-        controller.helpUrgentImage.setImage(new Image(Ecdar.class.getResource("ic_help_urgent.png").toExternalForm()));
-        fitSizeWhenAvailable(controller.helpUrgentImage, controller.helpUrgentPane);
-
-        controller.helpInputImage.setImage(new Image(Ecdar.class.getResource("ic_help_input.png").toExternalForm()));
-        fitSizeWhenAvailable(controller.helpInputImage, controller.helpInputPane);
-
-        controller.helpOutputImage.setImage(new Image(Ecdar.class.getResource("ic_help_output.png").toExternalForm()));
-        fitSizeWhenAvailable(controller.helpOutputImage, controller.helpOutputPane);
     }
 
     public static void fitSizeWhenAvailable(final ImageView imageView, final StackPane pane) {
@@ -565,17 +387,7 @@ public class EcdarPresentation extends StackPane {
      * @return A Boolean Property that is true if the grid has been turned on and false if it is off
      */
     public BooleanProperty toggleGrid(){
-        return EcdarController.getActiveCanvasPresentation().toggleGridUi();
-    }
-
-    public void showSnackbarMessage(final String message) {
-        JFXSnackbarLayout content = new JFXSnackbarLayout(message);
-        controller.snackbar.enqueue(new JFXSnackbar.SnackbarEvent(content, new Duration(3000)));
-    }
-
-    public void showHelp() {
-        controller.dialogContainer.setVisible(true);
-        controller.dialog.show(controller.dialogContainer);
+        return MainController.getActiveCanvasPresentation().toggleGridUi();
     }
 
     public EcdarController getController() {
