+++ src/main/java/ecdar/utility/helpers/ZoomHelper.java
@@ -1,6 +1,7 @@
 package ecdar.utility.helpers;
 
 import ecdar.controllers.EcdarController;
+import ecdar.controllers.MainController;
 import ecdar.presentations.CanvasPresentation;
 import ecdar.presentations.Grid;
 import ecdar.presentations.ModelPresentation;
@@ -102,11 +103,11 @@ public class ZoomHelper {
      */
     public void zoomToFit() {
         if (active) {
-            if (EcdarController.getActiveCanvasPresentation().getController().activeComponentPresentation == null) {
+            if (MainController.getActiveCanvasPresentation().getController().activeComponentPresentation == null) {
                 resetZoom();
                 return;
             }
-            double newScale = Math.min(canvasPresentation.getWidth() / EcdarController.getActiveCanvasPresentation().getController().activeComponentPresentation.getWidth() - 0.1, canvasPresentation.getHeight() / EcdarController.getActiveCanvasPresentation().getController().activeComponentPresentation.getHeight() - 0.2); //0.1 for width and 0.2 for height added for margin
+            double newScale = Math.min(canvasPresentation.getWidth() / MainController.getActiveCanvasPresentation().getController().activeComponentPresentation.getWidth() - 0.1, canvasPresentation.getHeight() / MainController.getActiveCanvasPresentation().getController().activeComponentPresentation.getHeight() - 0.2); //0.1 for width and 0.2 for height added for margin
 
             //Scale canvas
             canvasPresentation.setScaleX(newScale);
@@ -129,13 +130,13 @@ public class ZoomHelper {
      */
     private void centerComponentAndUpdateGrid(double newScale){
         // Check added to avoid NullPointerException
-        if(EcdarController.getActiveCanvasPresentation().getController().activeComponentPresentation != null){
-            updateGrid(newScale, EcdarController.getActiveCanvasPresentation().getController().activeComponentPresentation);
-        } else if (EcdarController.getActiveCanvasPresentation().getController().getActiveModel() != null) {
+        if(MainController.getActiveCanvasPresentation().getController().activeComponentPresentation != null){
+            updateGrid(newScale, MainController.getActiveCanvasPresentation().getController().activeComponentPresentation);
+        } else if (MainController.getActiveCanvasPresentation().getController().getActiveModel() != null) {
             // The canvas is currently showing an EcdarSystem object, so wee need to find the SystemPresentation in order to center it on screen
             SystemPresentation systemPresentation = null;
 
-            for (Node node : EcdarController.getActiveCanvasPresentation().getController().root.getChildren()) {
+            for (Node node : MainController.getActiveCanvasPresentation().getController().root.getChildren()) {
                 if (node instanceof SystemPresentation) {
                     systemPresentation = (SystemPresentation) node;
                     break;
