+++ src/main/resources/ecdar/presentations/LeftSimPanePresentation.fxml
@@ -1,0 +1,34 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+<?import ecdar.presentations.TracePaneElementPresentation?>
+<?import ecdar.presentations.TransitionPaneElementPresentation?>
+
+<fx:root xmlns:fx="http://javafx.com/fxml/1"
+         xmlns="http://javafx.com/javafx/8.0.76-ea"
+         type="StackPane"
+         fx:id="root"
+         fx:controller="ecdar.controllers.LeftSimPaneController"
+         minWidth="300">
+
+    <AnchorPane>
+        <ScrollPane fx:id="scrollPane"
+                    fitToHeight="true" fitToWidth="true"
+                    AnchorPane.topAnchor="0"
+                    AnchorPane.bottomAnchor="0"
+                    AnchorPane.leftAnchor="0"
+                    AnchorPane.rightAnchor="0"
+                    styleClass="edge-to-edge" >
+            <VBox fx:id="scrollPaneVbox">
+                <VBox AnchorPane.topAnchor="0" AnchorPane.leftAnchor="0" AnchorPane.rightAnchor="0">
+                    <TransitionPaneElementPresentation fx:id="transitionPanePresentation" />
+                </VBox>
+                <VBox AnchorPane.leftAnchor="0" AnchorPane.rightAnchor="0">
+                    <TracePaneElementPresentation fx:id="tracePanePresentation" />
+                </VBox>
+            </VBox>
+        </ScrollPane>
+    </AnchorPane>
+
+</fx:root>
