+++ src/main/resources/ecdar/presentations/RightSimPanePresentation.fxml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+
+<?import ecdar.presentations.QueryPanePresentation?>
+<fx:root xmlns:fx="http://javafx.com/fxml/1"
+         xmlns="http://javafx.com/javafx/8.0.76-ea"
+         type="StackPane"
+         fx:id="root"
+         fx:controller="ecdar.controllers.RightSimPaneController"
+         minWidth="400">
+
+    <AnchorPane>
+        <ScrollPane fitToHeight="true" fitToWidth="true"
+                    AnchorPane.topAnchor="0"
+                    AnchorPane.bottomAnchor="0"
+                    AnchorPane.leftAnchor="0"
+                    AnchorPane.rightAnchor="0"
+                    styleClass="edge-to-edge">
+            <VBox fx:id="scrollPaneVbox" >
+                <QueryPanePresentation fx:id="queryPane"/>
+            </VBox>
+        </ScrollPane>
+    </AnchorPane>
+</fx:root>
