+++ src/main/resources/ecdar/presentations/TracePaneElementPresentation.fxml
@@ -1,0 +1,49 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+
+<?import com.jfoenix.controls.JFXRippler?>
+<?import org.kordamp.ikonli.javafx.FontIcon?>
+<fx:root xmlns="http://javafx.com/javafx"
+            xmlns:fx="http://javafx.com/fxml"
+            type="AnchorPane"
+            fx:id="root"
+            fx:controller="ecdar.controllers.TracePaneElementController">
+
+    <VBox fx:id="traceVbox" AnchorPane.leftAnchor="0" AnchorPane.rightAnchor="0">
+        <AnchorPane fx:id="toolbar" minHeight="56" maxHeight="56">
+            <JFXRippler fx:id="expandTrace"
+                        AnchorPane.leftAnchor="8" AnchorPane.bottomAnchor="0" AnchorPane.topAnchor="0"
+                        onMouseClicked="#expandTrace">
+                <StackPane minWidth="40" minHeight="40">
+                    <FontIcon fx:id="expandTraceIcon" iconSize="24" fill="white"/>
+                </StackPane>
+            </JFXRippler>
+
+            <Label fx:id="traceTitle"
+                   styleClass="title"
+                   text="Trace" AnchorPane.leftAnchor="56" AnchorPane.topAnchor="0" AnchorPane.bottomAnchor="0"/>
+
+        </AnchorPane>
+
+        <VBox fx:id="traceList">
+            <!-- Trace goes here -->
+        </VBox>
+
+        <AnchorPane fx:id="traceSummary" minHeight="56" maxHeight="56" onMouseClicked="#expandTrace">
+            <VBox AnchorPane.leftAnchor="8"
+                  AnchorPane.topAnchor="0"
+                  AnchorPane.bottomAnchor="0"
+                  AnchorPane.rightAnchor="8"
+                  alignment="CENTER_LEFT">
+                <Label fx:id="summaryTitleLabel"
+                       text="0 steps in trace"
+                       styleClass="subhead"/>
+                <Label fx:id="summarySubtitleLabel"
+                       text="Expand to show trace..."
+                       styleClass="caption"/>
+            </VBox>
+        </AnchorPane>
+    </VBox>
+</fx:root>
