+++ src/main/resources/ecdar/presentations/TransitionPaneElementPresentation.fxml
@@ -1,0 +1,66 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+
+<?import com.jfoenix.controls.JFXRippler?>
+<?import org.kordamp.ikonli.javafx.FontIcon?>
+<?import com.jfoenix.controls.JFXTextField?>
+<fx:root xmlns="http://javafx.com/javafx"
+         xmlns:fx="http://javafx.com/fxml"
+         fx:controller="ecdar.controllers.TransitionPaneElementController"
+         fx:id="root"
+         type="AnchorPane">
+    <VBox fx:id="paneElementVbox" AnchorPane.leftAnchor="0" AnchorPane.rightAnchor="0">
+        <AnchorPane fx:id="toolbar"
+                    minHeight="56" maxHeight="56">
+            <JFXRippler fx:id="expandTransition"
+                        AnchorPane.leftAnchor="8" AnchorPane.bottomAnchor="0" AnchorPane.topAnchor="0"
+                        onMouseClicked="#expandTransitions">
+                <StackPane minWidth="40" minHeight="40">
+                    <FontIcon fx:id="expandTransitionIcon" iconSize="24" fill="white"/>
+                </StackPane>
+            </JFXRippler>
+
+            <Label fx:id="toolbarTitle"
+                   styleClass="title"
+                   text="Transitions" AnchorPane.leftAnchor="56" AnchorPane.topAnchor="0" AnchorPane.bottomAnchor="0"/>
+
+            <JFXRippler fx:id="refreshRippler" AnchorPane.rightAnchor="8" AnchorPane.topAnchor="0" AnchorPane.bottomAnchor="0"
+                        onMouseClicked="#refreshTransitions">
+                <StackPane minWidth="40" minHeight="40">
+                    <FontIcon fx:id="refreshIcon" iconLiteral="gmi-refresh" iconSize="24" fill="white"/>
+                </StackPane>
+            </JFXRippler>
+
+        </AnchorPane>
+        <AnchorPane fx:id="delayChooser">
+            <HBox AnchorPane.topAnchor="8"
+                  AnchorPane.bottomAnchor="8"
+                  AnchorPane.rightAnchor="8"
+                  AnchorPane.leftAnchor="8">
+                <VBox alignment="CENTER">
+                    <Label text="Delay"
+                           AnchorPane.rightAnchor="8"
+                           AnchorPane.leftAnchor="8"
+                           styleClass="subhead" />
+                </VBox>
+
+            </HBox>
+            <HBox AnchorPane.topAnchor="8"
+                  AnchorPane.bottomAnchor="8"
+                  AnchorPane.rightAnchor="0"
+                  AnchorPane.leftAnchor="0" alignment="CENTER" >
+                <JFXTextField fx:id="delayTextField"
+                              text="0.0"
+                              prefWidth="100"
+                              alignment="CENTER"
+                              styleClass="subhead"/>
+            </HBox>
+        </AnchorPane>
+
+        <VBox fx:id="transitionList">
+        <!-- Transitions goes here -->
+        </VBox>
+    </VBox>
+</fx:root>
